% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_management.R
\name{ammend.httk.option.list}
\alias{ammend.httk.option.list}
\title{Ammend an httk option list}
\usage{
ammend.httk.option.list(httk.option.list = NULL)
}
\arguments{
\item{httk.option.list}{An httk option that is formatted as a list may be: \code{Caco2.options}}
}
\value{
\item{httk.option.list}{The ammended list of options}
}
\description{
This function adds default values to httk options that are formatted as a list
in the case where the user neglected to include all of the options in the list.
}
\author{
Gregory Honda
}
