% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_mass_functions.R
\name{brain_mass}
\alias{brain_mass}
\title{Predict brain mass.}
\usage{
brain_mass(gender, age_years)
}
\arguments{
\item{gender}{Vector of genders, either 'Male' or 'Female'}

\item{age_years}{Vector of ages in years.}
}
\value{
A vector of brain masses in kg.
}
\description{
Predict brain mass from gender and age.
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to
environmental chemicals by simulating toxicokinetic variability."
Environment International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{httk-pop}
