% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_analytic_css_pbtk.R
\name{calc_analytic_css_pbtk}
\alias{calc_analytic_css_pbtk}
\title{Calculate the analytic steady state concentration for model pbtk.}
\usage{
calc_analytic_css_pbtk(chem.name = NULL, chem.cas = NULL,
  parameters = NULL, hourly.dose = 1/24, concentration = "plasma",
  suppress.messages = F, recalc.blood2plasma = F, tissue = NULL,
  restrictive.clearance = T, ...)
}
\arguments{
\item{chem.name}{Either the chemical name, CAS number, or the parameters must 
be specified.}

\item{chem.cas}{Either the chemical name, CAS number, or the parameters must 
be specified.}

\item{parameters}{Chemical parameters from parameterize_pbtk (for model = 
'pbtk'), parameterize_3comp (for model = '3compartment), 
parmeterize_1comp(for model = '1compartment') or parameterize_steadystate 
(for model = '3compartmentss'), overrides chem.name and chem.cas.}

\item{hourly.dose}{Hourly dose rate mg/kg BW/h.}

\item{concentration}{Desired concentration type, 'blood' or default 'plasma'.}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{recalc.blood2plasma}{Recalculates the ratio of the amount of chemical 
in the blood to plasma using the input parameters. Use this if you have 
'altered hematocrit, Funbound.plasma, or Krbc2pu.}

\item{tissue}{Desired tissue conentration (defaults to whole body 
concentration.)}

\item{restrictive.clearance}{If TRUE (default), then only the fraction of
chemical not bound to protein is available for metabolism in the liver. If 
FALSE, then all chemical in the liver is metabolized (faster metabolism due
to rapid off-binding).}

\item{...}{Additional parameters passed to parameterize function if 
parameters is NULL.}
}
\value{
Steady state concentration in uM units
}
\description{
This function calculates the analytic steady state plasma or venous blood 
concentrations as a result of infusion dosing.
}
\author{
Robert Pearce and John Wambaugh
}
