% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_oral_bioavailability.R
\name{calc_fbio.oral}
\alias{calc_fbio.oral}
\title{Parameterize_SteadyState}
\usage{
calc_fbio.oral(Params = NULL, chem.cas = NULL, chem.name = NULL,
  species = "Human", clint.pvalue.threshold = 0.05,
  default.to.human = F, human.clint.fup = F,
  adjusted.Funbound.plasma = T, restrictive.clearance = T,
  fup.lod.default = 0.005, suppress.messages = F,
  minimum.Funbound.plasma = 1e-04,
  Caco2.options = list(Caco2.Pab.default = 2, Caco2.Fgut = TRUE,
  Caco2.Fabs = TRUE))
}
\arguments{
\item{Params}{A list of the parameters (Caco2.Pab, Funbound.Plasma, Rblood2plasma,
cl_us, BW, Qsmallintestine, Fgutabs) used in the calculation, either supplied by user
or calculated in parameterize_steady_state.}

\item{chem.cas}{Either the chemical name or the CAS number must be
specified.}

\item{chem.name}{Either the chemical name or the CAS number must be
specified.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{clint.pvalue.threshold}{Hepatic clearances with clearance assays
having p-values greater than the threshold are set to zero.}

\item{default.to.human}{Substitutes missing rat values with human values if
true.}

\item{human.clint.fup}{Uses human hepatic intrinsic clearance and fraction
of unbound plasma in calculation of partition coefficients for rats if true.}

\item{adjusted.Funbound.plasma}{Returns adjusted Funbound.plasma when set to
TRUE.}

\item{restrictive.clearance}{In calculating hepatic.bioavailability, protein
binding is not taken into account (set to 1) in liver clearance if FALSE.}

\item{minimum.Funbound.plasma}{Monte Carlo draws less than this value are set 
equal to this value (default is 0.0001 -- half the lowest measured Fup in our
dataset).}

\item{Caco2.options}{A list of options to use when working with Caco2 apical to
basolateral data \item{Caco2.Pab}, default is Caco2.options = list(Caco2.default = 2,
Caco2.Fabs = TRUE, Caco2.Fgut = TRUE). Caco2.default sets the default value for 
Caco2.Pab if Caco2.Pab is unavailable. Caco2.Fabs = TRUE uses Caco2.Pab to calculate
fabs.oral, otherwise fabs.oral = \item {Fgutabs}. Caco2.Fgut = TRUE uses Caco2.Pab to calculate 
fgut.oral, otherwise fgut.oral = 1.}
}
\value{
\item{fbio.oral}{Oral bioavailability, the fraction of oral dose 
reaching systemic distribution in the body.} \item{fabs.oral}{Fraction of dose absorbed, 
i.e. the fraction of the dose that enters the gutlumen.} \item{fgut.oral}{Fraction of 
chemical surviving first pass metabolism in the gut.} \item{fhep.oral}{Fraction of chemical
surviving first pass hepatic clearance.}
}
\description{
These functions calculate the fraction absorbed (calc_fabs.oral), the fraction
surviving first pass gut metabolism (calc_fgut.oral), and the oral bioavailability
(calc_fbio.oral). Note that the first pass hepatic clearance is calculated within the
parameterization and other functions. Options are set in the list Caco2.options.
}
\examples{

 fbio1 <- calc_fbio.oral(chem.cas='80-05-7',
                         Caco2.options = list(Caco2.Pab.default = 2,
                                              Caco2.Fabs = TRUE,
                                              Caco2.Fgut = TRUE))
 fbio2 <- calc_fbio.oral(chem.cas='80-05-7',
                         Caco2.options = list(Caco2.Pab.default = 2,
                                              Caco2.Fabs = FALSE,
                                              Caco2.Fgut = FALSE))

}
\author{
Gregory Honda
}
\keyword{Parameter}
