% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_krbc2pu.R
\name{calc_krbc2pu}
\alias{calc_krbc2pu}
\title{Back-calculates the Red Blood Cell to Unbound Plasma Partition Coefficient}
\usage{
calc_krbc2pu(Rb2p, Funbound.plasma, hematocrit = NULL,
  default.to.human = F, species = "Human", suppress.messages = T)
}
\arguments{
\item{hematocrit}{Overwrites default hematocrit value in calculating
Rblood2plasma.}

\item{default.to.human}{Substitutes missing animal values with human values
if true.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{suppress.messages}{Determine whether to display certain usage
feedback.}

\item{chem.name}{Either the chemical name or the CAS number must be
specified.}

\item{chem.cas}{Either the CAS number or the chemical name must be
specified.}

\item{params}{Parameters from parameterize_schmitt.}

\item{adjusted.Funbound.plasma}{Whether or not to use Funbound.plasma
adjustment.}
}
\description{
Given and observed ratio of chemial concentration in blood to plasma, this
function calculates a Red Blood Cell to unbound plasma (Krbc2pu) partition
coefficient that would be consistent with that observation.
}
\references{
Pearce, Robert G., et al. "Evaluation and calibration of high-throughput 
predictions of chemical distribution to tissues." Journal of 
pharmacokinetics and pharmacodynamics 44.6 (2017): 549-565.

Ruark, Christopher D., et al. "Predicting passive and active tissue: plasma 
partition coefficients: interindividual and interspecies variability." 
Journal of pharmaceutical sciences 103.7 (2014): 2189-2198.
}
\author{
John Wambaugh and Robert Pearce
}
\keyword{Parameter}
