\name{chem.invivo.PK.aggregate.data}
\docType{data}
\alias{chem.invivo.PK.aggregate.data}
\title{Parameter Estimates from Wambaugh et al. (2018)}
\description{
This table includes 1 and 2 compartment fits of plasma concentration vs time data aggregated from chem.invivo.PK.data, performed in Wambaugh et al. 2018. Data includes volume of distribution (Vdist, L/kg), elimination rate (kelim, 1/h), gut absorption rate (kgutabs, 1/h), fraction absorbed (Fgutabs), and steady state concentration (Css, mg/L).}
\usage{chem.invivo.PK.aggregate.data}
\format{data.frame}
\source{Wambaugh et al. 2018 Toxicological Sciences, in press}
\author{
John Wambaugh
}
\keyword{ datasets }
