% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chem.lists}
\alias{chem.lists}
\title{Chemical membership in different research projects}
\format{A list containing ten lists.}
\usage{
chem.lists
}
\description{
A static list of lists identifying chemical membership in different research
projects. While it is our intent to keep these lists up-to-date, the
information here is only for convenience and should not be considered to be
definitive.
}
\references{
Bucher, J. R. (2008). Guest Editorial: NTP: New Initiatives, New
Alignment. Environ Health Perspect 116(1).

Judson, R. S., Houck, K. A., Kavlock, R. J., Knudsen, T. B., Martin, M. T.,
Mortensen, H. M., Reif, D. M., Rotroff, D. M., Shah, I., Richard, A. M. and
Dix, D. J. (2010). In Vitro Screening of Environmental Chemicals for
Targeted Testing Prioritization: The ToxCast Project. Environmental Health
Perspectives 118(4), 485-492.

Wambaugh, J. F., Wang, A., Dionisio, K. L., Frame, A., Egeghy, P., Judson,
R. and Setzer, R. W. (2014). High Throughput Heuristics for Prioritizing
Human Exposure to Environmental Chemicals. Environmental Science &
Technology, 10.1021/es503583j.

CDC (2014). National Health and Nutrition Examination Survey. Available at:
http://www.cdc.gov/nchs/nhanes.htm.
}
\author{
John Wambaugh
}
\keyword{data}
