% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_httkpop_1comp.R
\name{convert_httkpop_1comp}
\alias{convert_httkpop_1comp}
\title{Converts HTTK-Pop physiology into parameters relevant to the one
compartment model}
\usage{
convert_httkpop_1comp(httk.pop.biomets, chem.cas = NULL,
  chem.name = NULL, dtxsid = NULL, parameters = NULL)
}
\arguments{
\item{httk.pop.biomets}{A data.table containing the physiological
parameters as expected by HTTK (from \code{\link{httkpop_bio}}) and
\code{Funbound.plasma} and \code{Clint} values (from
\code{\link{draw_fup_clint}}).}

\item{chem.cas}{Chemical Abstract Services Registry Number (CAS-RN) -- if
parameters is not specified then the chemical must be identified by either
CAS, name, or DTXISD}

\item{chem.name}{Chemical name (spaces and capitalization ignored) --  if
parameters is not specified then the chemical must be identified by either
CAS, name, or DTXISD}

\item{dtxsid}{EPA's 'DSSTox Structure ID (http://comptox.epa.gov/dashboard)
 -- if parameters is not specified then the chemical must be identified by 
either CAS, name, or DTXSIDs}

\item{parameters}{Parameters from the appropriate parameterization function
for the model indicated by argument model}

\item{model}{Which HTTK model to use.}

\item{adjusted.Funbound.plasma}{Uses adjusted Funbound.plasma when set to TRUE.}

\item{regression}{Whether or not to use the regressions in calculating partition 
coefficients.}

\item{well.stirred.correction}{Uses correction in calculation of hepatic clearance 
for well-stirred model if TRUE for hepatic.model well-stirred. This assumes 
clearance re/lative to amount unbound in whole blood instead of plasma, but 
converted to use with plasma concentration.}

\item{restrictive.clearance}{Protein binding not taken into account (set to 1) in 
liver clearance if FALSE.}

\item{concentration}{Blood, plasma, or tissue concentration.}

\item{clint.pvalue.threshold}{Hepatic clearance for chemicals where the in vitro 
clearance assay result has a p-values greater than the threshold are set to zero.}
}
\value{
A data.table whose columns are the parameters of the HTTK model
 specified in \code{model}.
}
\description{
Converts HTTK-Pop physiology into parameters relevant to the one
compartment model
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to 
environmental chemicals by simulating toxicokinetic variability." Environment 
International 106 (2017): 105-118
}
\author{
Caroline Ring, John Wambaugh, and Greg Honda
}
\keyword{1compartment}
\keyword{httk-pop}
