% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_httkpop_3compss.R
\name{convert_httkpop_3compss}
\alias{convert_httkpop_3compss}
\title{Converts HTTK-Pop physiology into parameters relevant to the three
compartment steady-state model}
\usage{
convert_httkpop_3compss(parameters.dt, httkpop.dt, ...)
}
\arguments{
\item{parameters.dt}{A data.table containing the physiological
parameters as expected by HTTK (from \code{\link{httkpop_bio}}) and
\code{Funbound.plasma} and \code{Clint} values (from
\code{\link{draw_fup_clint}}).}
}
\value{
A data.table whose columns are the parameters of the HTTK model
 specified in \code{model}.
}
\description{
Converts HTTK-Pop physiology into parameters relevant to the three
compartment steady-state model
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to 
environmental chemicals by simulating toxicokinetic variability." Environment 
International 106 (2017): 105-118
}
\author{
Caroline Ring, John Wambaugh, and Greg Honda
}
\keyword{'}
\keyword{3oompss}
\keyword{httk-pop}
\keyword{monte-carlo}
