% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chem_id.R
\name{get_chem_id}
\alias{get_chem_id}
\title{Retrieve chemical identity from HTTK package}
\usage{
get_chem_id(chem.cas = NULL, chem.name = NULL, dtxsid = NULL)
}
\arguments{
\item{chem.cas}{CAS regstry number}

\item{chem.name}{Chemical name}

\item{dtxsid}{DSSTox Substance identifier}
}
\description{
Given one of chem.name, chem.cas (Chemical Abstract Service Registry Number),
or DTXSID (DSStox Substance Identifier \url{https://comptox.epa.gov/dashboard}) this
function checks if the chemical is available and, if so, returns all three
pieces of information.
}
\author{
John Wambaugh and Robert Pearce
}
\keyword{cheminformatics}
