% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invitroPK_param.R
\name{get_invitroPK_param}
\alias{get_invitroPK_param}
\title{Retrieve data from chem.physical_and_invitro.data table}
\usage{
get_invitroPK_param(param, species, chem.name = NULL, chem.cas = NULL,
  dtxsid = NULL)
}
\value{
The value of the parameter, if found
}
\description{
or fraction unbound in plasma) from the main HTTK data. This function looks
for species-specific values.
}
\author{
John Wambaugh and Robert Pearce
}
