% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/httk-package.R
\docType{data}
\name{howgate}
\alias{howgate}
\title{Howgate 2006}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 24 rows and 11 columns.}
\usage{
howgate
}
\description{
This data set is only used in Vignette 5.

This data set is only used in Vignette 5.
}
\keyword{data}
\keyword{datasets}
