% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitro_mc.R
\name{invitro_mc}
\alias{invitro_mc}
\title{Draw in vitro TK parameters including uncertainty and variability.}
\usage{
<<<<<<< HEAD:man/draw_fup_clint.Rd
draw_fup_clint(this.chem = NULL, parameters = NULL, nsamp,
  fup.meas.cv = 0.4, clint.meas.cv = 0.3, caco2.meas.sd = 0.3,
  fup.pop.cv = 0.3, clint.pop.cv = 0.3, caco2.pop.sd = 0.3,
  poormetab = TRUE, fup.lod = 0.01, fup.censored.dist = FALSE,
  adjusted.Funbound.plasma = T, clint.pvalue.threshold = 0.05,
  minimum.Funbound.plasma = 1e-04,
  Caco2.options = list(Caco2.Pab.default = "1.6", Caco2.Fgut = TRUE,
  Caco2.Fabs = TRUE, overwrite.invivo = FALSE, keepit100 = FALSE))
=======
invitro_mc(parameters.dt = NULL, samples, fup.meas.cv = 0.4,
  clint.meas.cv = 0.3, fup.pop.cv = 0.3, clint.pop.cv = 0.3,
  poormetab = TRUE, fup.lod = 0.01, fup.censored.dist = FALSE,
  adjusted.Funbound.plasma = T, clint.pvalue.threshold = 0.05,
  minimum.Funbound.plasma = 1e-04)
>>>>>>> cd6935617acdc1f8696861a41ecfb6190cbebda1:man/invitro_mc.Rd
}
\arguments{
\item{samples}{The number of samples to draw.}

\item{fup.meas.cv}{Coefficient of variation of distribution of measured
\code{Funbound.plasma} values.}

\item{clint.meas.cv}{Coefficient of variation of distribution of measured 
\code{Clint} values.}

\item{fup.pop.cv}{Coefficient of variation of distribution of population
\code{Funbound.plasma} values.}

\item{clint.pop.cv}{Coefficient of variation of distribution of population
\code{Clint} values.}

\item{poormetab}{Logical. Whether to include poor metabolizers in the Clint
distribution or not.}

\item{fup.lod}{The average limit of detection for \code{Funbound.plasma}, below
which distribution will be censored if fup.censored.dist is TRUE. Default 0.01.}

\item{fup.censored.dist}{Logical. Whether to draw \code{Funbound.plasma} from a
censored distribution or not.}

\item{adjusted.Funbound.plasma}{Uses adjusted Funbound.plasma when set to
TRUE.}

\item{clint.pvalue.threshold}{Hepatic clearance for chemicals where the in
vitro clearance assay result has a p-values greater than the threshold are
set to zero.}

\item{minimum.Funbound.plasma}{Monte Carlo draws less than this value are set 
equal to this value (default is 0.0001 -- half the lowest measured Fup in our
dataset).}

<<<<<<< HEAD:man/draw_fup_clint.Rd
\item{Caco2.options}{A list of options to use when working with Caco2 apical to
basolateral data \code{Caco2.Pab}, default is Caco2.options = list(Caco2.default = 2,
Caco2.Fabs = TRUE, Caco2.Fgut = TRUE, overwrite.invivo = FALSE, keepit100 = FALSE). Caco2.default sets the default value for 
Caco2.Pab if Caco2.Pab is unavailable. Caco2.Fabs = TRUE uses Caco2.Pab to calculate
fabs.oral, otherwise fabs.oral = \code{Fabs}. Caco2.Fgut = TRUE uses Caco2.Pab to calculate 
fgut.oral, otherwise fgut.oral = \code{Fgut}. overwrite.invivo = TRUE overwrites Fabs and Fgut in vivo values from literature with 
Caco2 derived values if available. keepit100 = TRUE overwrites Fabs and Fgut with 1 (i.e. 100 percent) regardless of other settings.}
=======
\item{this.chem}{The CAS number of one of the HTTK chemicals (see
\code{\link[httk]{get_cheminfo}}).}

\item{parameters}{A list of chemical-specific model parameters containing at
least Funbound.plasma, Clint, and Fhep.assay.correction.}
>>>>>>> cd6935617acdc1f8696861a41ecfb6190cbebda1:man/invitro_mc.Rd
}
\value{
A data.table with three columns: \code{Funbound.plasma} and
\code{Clint}, containing the sampled values, and
\code{Fhep.assay.correction}, containing the value for fraction unbound in
hepatocyte assay.
}
\description{
Given a CAS in the HTTK data set, a virtual population from HTTK-Pop, some
user specifications on the assumed distributions of Funbound.plasma and
Clint, draw "individual" values of Funbound.plasma and Clint from those
distributions.
}
\references{
Wambaugh, John F., et al. "Assessing Toxicokinetic Uncertainty and 
Variability in Risk Prioritization." Toxicological Sciences (2019).
}
\author{
Caroline Ring and John Wambaugh
}
\keyword{in-vitro}
\keyword{monte-carlo}
