% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{johnson}
\alias{johnson}
\title{Johnson 2006}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 60 rows and 11 columns.}
\usage{
johnson
}
\description{
This data set is only used in Vignette 5.
}
\keyword{data}
