% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_mass_functions.R
\name{lung_mass_children}
\alias{lung_mass_children}
\title{Predict lung mass for children.}
\usage{
lung_mass_children(height, weight, gender)
}
\arguments{
\item{height}{Vector of heights in cm.}

\item{weight}{Vector of weights in kg.}

\item{gender}{Vector of genders (either 'Male' or 'Female').}
}
\value{
A vector of lung masses in kg.
}
\description{
For individuals under 18, predict the liver mass from height, weight, and
gender, using equations from Ogiu et al.
}
