% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabolism_data_Linakis2020}
\alias{metabolism_data_Linakis2020}
\title{Metabolism data involved in Linakis 2020 vignette analysis.}
\format{A data.frame containing x rows and y columns.}
\source{
Matt Linakis
}
\usage{
metabolism_data_Linakis2020
}
\description{
Metabolism data involved in Linakis 2020 vignette analysis.
}
\references{
DSStox database (http:// www.epa.gov/ncct/dsstox
}
\author{
Matt Linakis
}
\keyword{data}
