\name{Obach2008}
\docType{data}
\alias{Obach2008}
\title{Published Pharmacokinetic Parameters from Obach et al. 2008}
\description{
This data set is used in Vignette 4 for steady state concentration.
}
\usage{Obach2008}
\format{A data.frame containing 670 rows and 8 columns.}

\references{
Obach, R. Scott, Franco Lombardo, and Nigel J. Waters. "Trend analysis of a database of intravenous pharmacokinetic parameters in humans for 670 drug compounds." Drug Metabolism and Disposition 36.7 (2008): 1385-1405.
}
\keyword{datasets}
