% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameterize_SteadyState.R
\name{parameterize_steadystate}
\alias{parameterize_steadystate}
\title{Parameterize_SteadyState}
\usage{
parameterize_steadystate(chem.cas = NULL, chem.name = NULL,
  species = "Human", clint.pvalue.threshold = 0.05,
  default.to.human = F, human.clint.fup = F,
  adjusted.Funbound.plasma = T, restrictive.clearance = T,
  fup.lod.default = 0.005, suppress.messages = F,
  minimum.Funbound.plasma = 1e-04,
  Caco2.options = list(Caco2.Pab.default = 2, Caco2.Fgut = TRUE,
  Caco2.Fabs = TRUE))
}
\arguments{
\item{chem.cas}{Either the chemical name or the CAS number must be
specified.}

\item{chem.name}{Either the chemical name or the CAS number must be
specified.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{clint.pvalue.threshold}{Hepatic clearances with clearance assays
having p-values greater than the threshold are set to zero.}

\item{default.to.human}{Substitutes missing rat values with human values if
true.}

\item{human.clint.fup}{Uses human hepatic intrinsic clearance and fraction
of unbound plasma in calculation of partition coefficients for rats if true.}

\item{adjusted.Funbound.plasma}{Returns adjusted Funbound.plasma when set to
TRUE.}

\item{restrictive.clearance}{In calculating hepatic.bioavailability, protein
binding is not taken into account (set to 1) in liver clearance if FALSE.}

\item{fup.lod.default}{Default value used for fraction of unbound plasma for
chemicals where measured value was below the limit of detection. Default
value is 0.0005.}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{minimum.Funbound.plasma}{Monte Carlo draws less than this value are set 
equal to this value (default is 0.0001 -- half the lowest measured Fup in our
dataset).}

\item{Caco2.options}{A list of options to use when working with Caco2 apical to
basolateral data \item{Caco2.Pab}, default is Caco2.options = list(Caco2.default = 2,
Caco2.Fabs = TRUE, Caco2.Fgut = TRUE). Caco2.default sets the default value for 
Caco2.Pab if Caco2.Pab is unavailable. Caco2.Fabs = TRUE uses Caco2.Pab to calculate
fabs.oral, otherwise fabs.oral = \item {Fgutabs}. Caco2.Fgut = TRUE uses Caco2.Pab to calculate 
fgut.oral, otherwise fgut.oral = 1.}
}
\value{
\item{Clint}{Hepatic Intrinsic Clearance, uL/min/10^6 cells.}
\item{Fgutabs}{Fraction of the oral dose absorbed, i.e. the fraction of the
dose that enters the gutlumen.} \item{Funbound.plasma}{Fraction of plasma
that is not bound.} \item{Qtotal.liverc}{Flow rate of blood exiting the
liver, L/h/kg BW^3/4.} \item{Qgfrc}{Glomerular Filtration Rate, L/h/kg
BW^3/4, volume of fluid filtered from kidney and excreted.} \item{BW}{Body
Weight, kg} \item{MW}{Molecular Weight, g/mol}
\item{million.cells.per.gliver}{Millions cells per gram of liver tissue.}
\item{Vliverc}{Volume of the liver per kg body weight, L/kg BW.}
\item{liver.density}{Liver tissue density, kg/L.}
\item{Fhep.assay.correction}{The fraction of chemical unbound in hepatocyte
assay using the method of Kilford et al. (2008)}
\item{hepatic.bioavailability}{Fraction of dose remaining after first pass
clearance, calculated from the corrected well-stirred model.}
}
\description{
This function initializes the parameters needed in the functions
calc_mc_css, calc_mc_oral_equiv, and calc_analytic_css for the three
compartment steady state model ('3compartmentss').
}
\examples{

 parameters <- parameterize_steadystate(chem.name='Bisphenol-A',species='Rat')
 parameters <- parameterize_steadystate(chem.cas='80-05-7')

}
\author{
John Wambaugh
}
\keyword{Parameter}
