% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate_invitrouv_pbtk.R
\name{propagate_invitrouv_pbtk}
\alias{propagate_invitrouv_pbtk}
\title{Propagates uncertainty and variability in in vitro HTTK data into PBPK
model parameters}
\usage{
propagate_invitrouv_pbtk(parameters.dt, ...)
}
\arguments{
\item{parameters.dt}{The data table of parameters being used by the Monte
Carlo sampler}

\item{...}{Additional arguments passed to \code{\link{calc_hep_clearance}}}
}
\value{
A data.table whose columns are the parameters of the HTTK model
 specified in \code{model}.
}
\description{
Propagates uncertainty and variability in in vitro HTTK data into PBPK
model parameters
}
\author{
John Wambaugh
}
\keyword{monte-carlo}
\keyword{pbtk}
