% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagateuv_3comp.R
\name{propagateuv_3comp}
\alias{propagateuv_3comp}
\title{Propagates uncertainty and variability in in vitro HTTK data into three
compartment model parameters}
\usage{
propagateuv_3comp(parameters.dt, ...)
}
\arguments{
\item{parameters.dt}{The data table of parameters being used by the Monte
Carlo sampler}
}
\value{
A data.table whose columns are the parameters of the HTTK model
 specified in \code{model}.
}
\description{
Propagates uncertainty and variability in in vitro HTTK data into three
compartment model parameters
}
\author{
John Wambaugh
}
\keyword{3compartment}
\keyword{monte-carlo}
