% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_httk.R
\name{reset_httk}
\alias{reset_httk}
\title{Reset HTTK to Default Data Tables}
\usage{
reset_httk(target.env = .GlobalEnv)
}
\arguments{
\item{target.env}{The environment where the new
chem.physical_and_invitro.data is loaded. Defaults to global environment.}
}
\value{
\item{data.frame}{The package default version of 
chem.physical_and_invitro.data.}
}
\description{
This function returns an updated version of chem.physical_and_invitro.data
that includes data predicted with Simulations Plus' ADMET predictor that was
used in Sipes et al. 2017, included in admet.data.
}
\examples{

\dontrun{
chem.physical_and_invitro.data <- load_sipes2017()
reset_httk()
}                        

}
\author{
John Wambaugh
}
