% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rfun}
\alias{rfun}
\title{Randomly draws from a one-dimensional KDE}
\usage{
rfun(n, fhat)
}
\arguments{
\item{n}{Number of samples to draw}

\item{fhat}{A list with elements x, w, and h (h is the KDE bandwidth).}
}
\value{
A vector of n samples from the KDE fhat
}
\description{
Randomly draws from a one-dimensional KDE
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to
environmental chemicals by simulating toxicokinetic variability."
Environment International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{httk-pop}
