% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_dosing.R
\name{scale_dosing}
\alias{scale_dosing}
\title{Scale mg/kg body weight doses according to body weight and units}
\usage{
scale_dosing(dosing, parameters, route, output.units = "uM")
}
\description{
This function transforms the dose (in mg/kg) into the appropriate units. It
handles single doses, matrices of doses, or daily repeated doses (at varying
intervals. Gut absorption is also factored in through the parameter Fgutabs,
and scaling is currently avoided in the inhalation exposure case with a 
scale factor of 1
}
\author{
John Wambaugh
}
\keyword{Dynamic}
