% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_httk_precision.R
\name{set_httk_precision}
\alias{set_httk_precision}
\title{set_httk_precision}
\usage{
set_httk_precision(in.num, sig.fig = 4, num.prec = 9)
}
\arguments{
\item{in.num}{The numeric variabile (or assembly of numerics) to be
processd.}

\item{sig.fig}{The number of significant figures reported. Defaults to 4.}

\item{num.prec}{The precision maintained, digits below 10^num.prec are 
dropped. Defaults to 9.}
}
\value{
numeric values
}
\description{
Although the ODE solver and other functions return very precise numbers,
we cannot (or at least do not spend enough computing time to) be sure of the 
precioion to an arbitrary level. This function both limits the number of
signficant figures reported and truncates the numerical precision.
}
\author{
John Wambaugh
}
