% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_mass_functions.R
\name{skin_mass_bosgra}
\alias{skin_mass_bosgra}
\title{Predict skin mass.}
\usage{
skin_mass_bosgra(BSA)
}
\arguments{
\item{BSA}{Vector of body surface areas in cm^2.}
}
\value{
Vector of skin masses in kg.
}
\description{
Using equation from Bosgra et al. 2012, predict skin mass from body surface area.
}
