% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_fetal_pbtk.R
\name{solve_fetal_pbtk}
\alias{solve_fetal_pbtk}
\title{Solve_fetal_PBTK}
\usage{
solve_fetal_pbtk(chem.name = NULL, chem.cas = NULL, times = NULL,
  parameters = NULL, days = 10, tsteps = 4, daily.dose = 1,
  dose = NULL, doses.per.day = NULL, initial.values = NULL,
  plots = F, suppress.messages = F, species = "Human", iv.dose = F,
  output.units = "uM", method = "lsoda", rtol = 1e-08,
  atol = 1e-12, begin.css = F, default.to.human = F,
  recalc.blood2plasma = F, recalc.clearance = F,
  dosing.matrix = NULL, adjusted.Funbound.plasma = T, regression = T,
  restrictive.clearance = T, minimum.Funbound.plasma = 1e-04, ...)
}
\arguments{
\item{chem.name}{Either the chemical name, CAS number, or the parameters
must be specified.}

\item{chem.cas}{Either the chemical name, CAS number, or the parameters must
be specified.}

\item{times}{Optional time sequence for specified number of days.  Dosing
sequence begins at the beginning of times.}

\item{parameters}{Chemical parameters from parameterize_fetal
_pbtk function,
overrides chem.name and chem.cas.}

\item{days}{Length of the simulation.}

\item{tsteps}{The number time steps per hour.}

\item{daily.dose}{Total daily dose, mg.}

\item{dose}{Amount of a single dose, mg.  Overwrites daily.dose.}

\item{doses.per.day}{Number of doses per day.}

\item{initial.values}{Vector containing the initial concentrations or
amounts of the chemical in specified tissues with units corresponding to
output.units.  Defaults are zero.}

\item{plots}{Plots all outputs if true.}

\item{suppress.messages}{Whether or not the output message is suppressed.}

\item{species}{Species desired (either "Rat", "Rabbit", "Dog", "Mouse", or
default "Human").}

\item{iv.dose}{Simulates a single i.v. dose if true.}

\item{output.units}{Desired units (either "mg/L", "mg", "umol", or default
"uM").}

\item{method}{Method used by integrator (deSolve).}

\item{rtol}{Argument passed to integrator (deSolve).}

\item{atol}{Argument passed to integrator (deSolve).}

\item{begin.css}{Begin at steady state concentration in mother.}

\item{default.to.human}{Substitutes missing animal values with human values
if true (hepatic intrinsic clearance or fraction of unbound plasma).}

\item{recalc.blood2plasma}{Recalculates the ratio of the amount of chemical
in the blood to plasma using the input parameters, calculated with
hematocrit, Funbound.plasma, and Krbc2pu.}

\item{recalc.clearance}{Recalculates the the hepatic clearance
(Clmetabolism) with new million.cells.per.gliver parameter.}

\item{dosing.matrix}{A matrix of either one column (or row) with a set of
dosing times or with two columns (or rows) correspondingly named "dose" and
"time" containing the time and amount, in mg/kg BW, of each dose.}

\item{adjusted.Funbound.plasma}{Uses adjusted Funbound.plasma when set to
TRUE along with partition coefficients calculated with this value.}

\item{regression}{Whether or not to use the regressions in calculating
partition coefficients.}

\item{restrictive.clearance}{Protein binding not taken into account (set to
1) in liver clearance if FALSE.}

\item{minimum.Funbound.plasma}{Monte Carlo draws less than this value are set 
equal to this value (default is 0.0001 -- half the lowest measured Fup in our
dataset).}

\item{...}{Additional arguments passed to the integrator.}
}
\value{
A matrix of class deSolve with a column for time(in days), each
compartment, the area under the curve, and plasma concentration and a row
for each time point.
}
\description{
This function solves for the amounts or concentrations in uM of a chemical
in different tissues as functions of time based on the dose and dosing
frequency.
}
\details{
Note that the model parameters have units of hours while the model output is
in days. Dose is in mg, not scaled for body weight.

Default NULL value for doses.per.day solves for a single dose.

The compartments used in this model are the gutlumen, gut, liver, kidneys,
veins, arteries, lungs, and the rest of the body.

The extra compartments include the amounts or concentrations metabolized by
the liver and excreted by the kidneys through the tubules.

AUC is the area under the curve of the plasma concentration.

When species is specified as rabbit, dog, or mouse, the function uses the
appropriate physiological data (volumes and flows) but substitues human
fraction unbound, partition coefficients, and intrinsic hepatic clearance.
}
\examples{



}
\author{
John Wambaugh and Robert Pearce
}
\keyword{Solve}
