% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_mass_functions.R
\name{spleen_mass_children}
\alias{spleen_mass_children}
\title{Predict spleen mass for children.}
\usage{
spleen_mass_children(height, weight, gender)
}
\arguments{
\item{height}{Vector of heights in cm.}

\item{weight}{Vector of weights in kg.}

\item{gender}{Vector of genders (either 'Male' or 'Female').}
}
\value{
A vector of spleen masses in kg.
}
\description{
For individuals under 18, predict the spleen mass from height, weight, and
gender, using equations from Ogiu et al.
}
\references{Ring, Caroline L., et al. "Identifying populations sensitive to environmental chemicals by simulating toxicokinetic variability." Environment International 106 (2017): 105-118.}
