% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_masses_flows.R
\name{tissue_masses_flows}
\alias{tissue_masses_flows}
\title{Given a data.table describing a virtual population by the NHANES quantities, 
generates HTTK physiological parameters for each individual.}
\usage{
tissue_masses_flows(tmf_dt)
}
\arguments{
\item{tmf_dt}{A data.table generated by
\code{gen_age_height_weight()}, containing variables \code{gender},
\code{reth}, \code{age_months}, \code{age_years}, \code{weight}, and
\code{height}.}
}
\value{
The same data.table, with aditional variables describing tissue masses
 and flows.
}
\description{
Given a data.table describing a virtual population by the NHANES quantities, 
generates HTTK physiological parameters for each individual.
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to 
environmental chemicals by simulating toxicokinetic variability." Environment 
International 106 (2017): 105-118
}
\author{
Caroline Ring
}
\keyword{httk-pop}
