% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wambaugh2019.Tox21}
\alias{wambaugh2019.Tox21}
\title{Subset of the ToxCast and Tox21 Bioactivites Observed with HTS}
\format{A data frame with 489 rows and 1197 variables:}
\source{
\url{ftp://newftp.epa.gov/COMPTOX/High_Throughput_Screening_Data/Previous_Data/ToxCast_Data_Release_Oct_2015/}
}
\usage{
wambaugh2019.Tox21
}
\description{
These data are all ToxCast and Tox21 high throughput screening (HTS) assay
results for the subset of chemicals appearing in Wambaugh et al. (2019). The
values are NA if no systematic concentration-response was observed, otherwise
the log base ten uM concentration of the activity concentration at cutoff (ACC)
is given. Each column correspons to a different in vitro HTS assay.
}
\references{
Wambaugh et al. "Assessing Toxicokinetic Uncertainty and 
Variability in Risk Prioritization", submitted.
}
\author{
John Wambaugh
}
\keyword{data}
