% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_httkpop_1comp.R
\name{convert_httkpop_1comp}
\alias{convert_httkpop_1comp}
\title{Converts HTTK-Pop physiology into parameters relevant to the one
compartment model}
\usage{
convert_httkpop_1comp(parameters.dt, httkpop.dt, ...)
}
\arguments{
\item{parameters.dt}{Data table returned by \code{\link{create_mc_samples}}}

\item{httkpop.dt}{Data table returned by \code{\link{httkpop_generate}}}

\item{...}{Additional arguments passed to \code{\link{propagate_invitrouv_1comp}}}
}
\value{
A data.table whose columns are the parameters of the HTTK model
 specified in \code{model}.
}
\description{
Converts HTTK-Pop physiology into parameters relevant to the one
compartment model
}
\references{
Ring, Caroline L., et al. "Identifying populations sensitive to 
environmental chemicals by simulating toxicokinetic variability." Environment 
International 106 (2017): 105-118
}
\author{
Caroline Ring, John Wambaugh, and Greg Honda
}
\keyword{1compartment}
\keyword{httk-pop}
