% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_dosing.R
\name{scale_dosing}
\alias{scale_dosing}
\title{Scale mg/kg body weight doses according to body weight and units}
\usage{
scale_dosing(dosing, parameters, route, output.units = "uM")
}
\arguments{
\item{dosing}{List of dosing metrics to be scaled, possibly including
"initial.dose", "dosing.matrix", and/or "daily.dose".}

\item{parameters}{List of chemical parameters, as output by 
parameterize_pbtk function. Overrides chem.name and chem.cas.}

\item{route}{String specification of route of exposure for simulation: 
"oral", "iv", ...}

\item{output.units}{String specification of desired output units, "uM",
"umol", "mg", or "mg/L"}
}
\value{
Modified dosing list scaled with entries scaled according to desired
output units, body weight, and route of exposure.
}
\description{
This function transforms the dose (in mg/kg) into the appropriate units. It
handles single doses, matrices of doses, or daily repeated doses (at varying
intervals. Gut absorption is also factored in through the parameter Fgutabs,
and scaling is currently avoided in the inhalation exposure case with a 
scale factor of 1
}
\author{
John Wambaugh and Mark Sfeir
}
\keyword{Dynamic}
