\name{Sipes2017}
\docType{data}
\alias{Sipes2017}
\title{Sipes et al. 2017 data}
\description{
This table includes data predicted with Simulations Plus' ADMET predictor, used in load_sipes2017, that was used in Sipes et al. 2017. The column names are equivalent to those of chem.physical_and_invitro.data.}
\usage{Sipes2017}
\format{data.frame}
\source{ADMET, Simulations Plus}
\references{
Sipes, Nisha S., et al. "An Intuitive Approach for Predicting Potential Human Health Risk with the Tox21 10k Library." Environmental Science & Technology 51.18 (2017): 10786-10796.
}
\author{
Nisha Sipes
}
\keyword{ datasets }
